fix(ct): keep contenttools package under TD.2 budget for CT.21
Inline Class Pulse metrics into the actions file, fold Parameter Explorer
metrics into its actions file, remove dead RowsFromStateBlobs, and register
CT.21 in the e2e completed-feature coverage manifest.
